pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "routing-data"
version = "0.1.0"
description = "TSPLIB95 ETL Converter - Convert TSPLIB/VRP instances to JSON and DuckDB"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "TCC Routing Data Project", email = "lucasgaldinos@example.com"}
]
license = {text = "MIT"}
keywords = ["tsp", "vrp", "routing", "optimization", "tsplib95", "etl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Mathematics",
]

dependencies = [
    "click>=8.3.0",
    "deprecated>=1.2.18",
    "duckdb>=1.4.0",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "rich>=14.1.0",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/lucasgaldinos/Routing_data"
Repository = "https://github.com/lucasgaldinos/Routing_data.git"
Issues = "https://github.com/lucasgaldinos/Routing_data/issues"

[project.scripts]
converter = "converter.cli.commands:cli"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
